v0.17
v0.17

+++ Assets/Shaders/Pal3Dev.shader
@@ -53,9 +53,9 @@ Shader "Pal3/Dev"
                     return o;
                 }
 
-                fixed4 frag (v2f i) : SV_Target
+                half4 frag (v2f i) : SV_Target
                 {
-                    fixed4 col = tex2D(_MainTex, i.texcoord);
+                    half4 col = tex2D(_MainTex, i.texcoord);
                     UNITY_APPLY_FOG(i.fogCoord, col);
                     col.a = 0.5;
                     return col;
