+++ Assets/Shaders/Pal3Standard.shader
@@ -68,9 +68,9 @@ Shader "Pal3/Standard"
                 return o;
             }
 
-            fixed4 frag(v2f i) : SV_Target
+            half4 frag(v2f i) : SV_Target
             {
-                fixed4 color = tex2D(_MainTex, i.texcoord);
+                half4 color = tex2D(_MainTex, i.texcoord);
 
                 // Cutout
                 clip(color.a - _Cutoff);
